refactor: move compiler, executor, etc. to the tools dir
